# services/engine/openapi.yaml
openapi: 3.0.3
info:
  title: Instance Service API
  description: API for managing compute instances
  version: 1.0.0

servers:
  - url: https://api.example.com

tags:
  - name: instances
    description: Instance lifecycle operations

paths:
  /instances:
    post:
      tags: [instances]
      summary: Create a new instance
      description: Inter-server endpoint for creating instances
      security:
        - InterserverSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstanceRequest"
      responses:
        "201":
          description: Instance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateInstanceResponse"
        "401":
          description: Unauthorized

    get:
      tags: [instances]
      summary: List user instances
      description: Returns instances owned by the authenticated user
      security:
        - BearerJWT: []
      responses:
        "200":
          description: List of instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstanceSummary"
        "401":
          description: Unauthorized
  /instances/{identifier}/lifetime:
    post:
      tags: [instances]
      summary: Extend instance lifetime
      description: Inter-server endpoint to update the expiration of an existing instance
      security:
        - InterserverSignature: []
      parameters:
        - $ref: "#/components/parameters/InstanceIdentifier"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstanceLifetimeRequest"
      responses:
        "200":
          description: Instance expiration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateInstanceLifetimeResponse"
        "401":
          description: Unauthorized
        "404":
          description: Instance not found
        "400":
          description: Invalid request

  /instances/{identifier}:
    get:
      tags: [instances]
      summary: Get instance status
      description: Returns status for a single instance
      security:
        - BearerJWT: []
      parameters:
        - $ref: "#/components/parameters/InstanceIdentifier"
      responses:
        "200":
          description: Instance status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceDetail"
        "401":
          description: Unauthorized
        "404":
          description: Instance not found

    put:
      tags: [instances]
      summary: Update instance specification
      description: Updates limited, user-configurable instance parameters
      security:
        - BearerJWT: []
      parameters:
        - $ref: "#/components/parameters/InstanceIdentifier"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstanceRequest"
      responses:
        "200":
          description: Instance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceDetail"
        "401":
          description: Unauthorized
        "404":
          description: Instance not found

    delete:
      tags: [instances]
      summary: Delete instance
      description: Inter-server endpoint for deleting an instance
      security:
        - InterserverSignature: []
      parameters:
        - $ref: "#/components/parameters/InstanceIdentifier"
      responses:
        "204":
          description: Instance deleted
        "401":
          description: Unauthorized
        "404":
          description: Instance not found

components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT

    InterserverSignature:
      type: apiKey
      in: header
      name: X-Signature

  parameters:
    InstanceIdentifier:
      name: identifier
      in: path
      required: true
      description: Unique instance identifier
      schema:
        type: string

  schemas:
    CreateInstanceRequest:
      type: object
      required:
        - image
        - owner_id
      properties:
        image:
          type: string
          description: Source image identifier
        owner_id:
          type: string
          description: Owner user identifier
        hardware:
          type: object
          description: Hardware specification
          properties:
            cpu:
              type: integer
              minimum: 1
            memory_mb:
              type: integer
              minimum: 128
            disk_gb:
              type: integer
              minimum: 1

    CreateInstanceResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Newly created instance identifier

    UpdateInstanceRequest:
      type: object
      properties:
        metadata:
          type: object
          additionalProperties:
            type: string
        configuration:
          type: object
          description: Allowed user-modifiable configuration options
          additionalProperties: true

    InstanceSummary:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/InstanceStatus"

    InstanceDetail:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/InstanceStatus"
        metadata:
          type: object
          additionalProperties:
            type: string
        configuration:
          type: object
          additionalProperties: true

    InstanceStatus:
      type: string
      enum:
        - creating
        - running
        - stopped
        - deleting
        - error

    UpdateInstanceLifetimeRequest:
      type: object
      required:
        - new_expiration
      properties:
        new_expiration:
          type: integer
          description: New expiration timestamp in Unix seconds

    UpdateInstanceLifetimeResponse:
      type: object
      required:
        - id
        - new_expiration
      properties:
        id:
          type: string
          description: Instance identifier
        new_expiration:
          type: integer
          description: Updated expiration timestamp in Unix seconds